+++ src/src/App.vue
@@ -1,10 +1,10 @@
 <template>
   <Header />
   <div class="container">
-    <Balance :total="total" />
+    <Balance :total="+total" />
     <IncomeExpenses :income="+income" :expenses="+expenses" />
     <TransactionList :transactions="transactions" />
-    <AddTransaction />
+    <AddTransaction @transactionSubmitted="handleTransactionSubmitted" />
   </div>
 </template>
 
@@ -48,4 +48,9 @@ const expenses = computed(() => {
       return acc + transaction.amount
   }, 0).toFixed(2);
 });
+
+// Add transaction
+const handleTransactionSubmitted = (transactionData) => {
+  console.log(transactionData);
+};
 </script>