+++ src/src/components/AddTransaction.vue
@@ -23,6 +23,7 @@ import { useToast } from 'vue-toastification';
 const text = ref();
 const amount = ref();
 
+const emit = defineEmits(['transactionSubmitted']);
 const toast = useToast();
 
 const onSubmit = () => {
@@ -30,7 +31,13 @@ const onSubmit = () => {
     toast.error('Both fields must be filled!');
     return;
   }
-  console.log(text.value, amount.value);
+  
+  const transactionData = {
+    text: text.value,
+    amount: parseFloat(amount.value)
+  }
+
+  emit('transactionSubmitted', transactionData);
 
   text.value = '';
   amount.value = '';