+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     ```bash
     sudo npm i vue-toastification@next
     ```
+    - add event emmited on form submit on add transaction
 
 - 2024-03-07
     - init